requirements.txt
langchain
PyPDF2
faiss-cpu
python-dotenv
google-generativeai
streamlit
langchain_google_genai